+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -198,6 +198,7 @@ import Cardano.Ledger.Keys (
  )
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.API.ByronTranslation (translateToShelleyLedgerStateFromUtxo)
+import Cardano.Ledger.Shelley.API.Mempool (ApplyTx (..))
 import Cardano.Ledger.Shelley.API.Validation (
   ApplyBlock,
   BlockTransitionError (..),
@@ -218,12 +219,14 @@ import Cardano.Ledger.Shelley.LedgerState (
   curPParamsEpochStateL,
   esLStateL,
   lsCertStateL,
+  lsUTxOState,
   lsUTxOStateL,
   nesELL,
   nesEsL,
   prevPParamsEpochStateL,
   produced,
   utxosDonationL,
+  utxosUtxo,
  )
 import Cardano.Ledger.Shelley.Rules (
   BbodyEnv (..),
@@ -234,6 +237,7 @@ import Cardano.Ledger.Shelley.Rules (
   ShelleyUtxoPredFailure,
   ShelleyUtxowPredFailure,
   epochFromSlot,
+  ledgerPpL,
  )
 import Cardano.Ledger.Shelley.Scripts (
   ShelleyEraScript,
@@ -441,6 +445,7 @@ impRecordedTxsL = lens impRecordedTxs (\x y -> x {impRecordedTxs = y})
 
 class
   ( ShelleyEraTest era
+  , ApplyTx era
   , -- For the BBODY rule
     STS (EraRule "BBODY" era)
   , BaseM (EraRule "BBODY" era) ~ ShelleyBase
@@ -463,7 +468,7 @@ class
   , -- For the LEDGER rule
     STS (EraRule "LEDGER" era)
   , BaseM (EraRule "LEDGER" era) ~ ShelleyBase
-  , Signal (EraRule "LEDGER" era) ~ Tx TopTx era
+  , Signal (EraRule "LEDGER" era) ~ StAnnTx TopTx era
   , State (EraRule "LEDGER" era) ~ LedgerState era
   , Environment (EraRule "LEDGER" era) ~ LedgerEnv era
   , Eq (PredicateFailure (EraRule "LEDGER" era))
@@ -1256,11 +1261,19 @@ trySubmitTx tx = do
 
   st <- gets impNES
   lEnv <- impLedgerEnv st
-  res <- tryRunImpRule @"LEDGER" lEnv (st ^. nesEsL . esLStateL) txFixed
+  globals <- use impGlobalsL
+  let lState = st ^. nesEsL . esLStateL
+      stAnnTx =
+        mkStAnnTx
+          (epochInfo globals)
+          (systemStart globals)
+          (lEnv ^. ledgerPpL)
+          (utxosUtxo (lsUTxOState lState))
+          txFixed
+  res <- tryRunImpRule @"LEDGER" lEnv lState stAnnTx
 
   -- Check for conformance
-  globals <- use impGlobalsL
-  let trc = TRC (lEnv, st ^. nesEsL . esLStateL, txFixed)
+  let trc = TRC (lEnv, lState, stAnnTx)
   asks itePostSubmitTxHook >>= (\f -> f globals trc res)
 
   case res of